skip hidden flows (non-linear fragments) in the progress bar
A user reported that when reading a book with hidden flows (non-linear
fragments), the progress bar on PocketBook home screen never reaches
100% - https://www.mobileread.com/forums/showpost.php?p=4371306&postcount=30

To solve this, instead of using the total number of pages, we use the
number of pages in the main flow (actually, current flow, and if that is
not the main one, we skip the sync completely).

The implementation is inspired by
https://github.com/koreader/koreader/blob/2e2ca76a03ba255764bb9b8e6cc81a4bd501a3b3/plugins/statistics.koplugin/main.lua#L1620-L1630
Note that we don't need to check hasHiddenFlows as the flow-aware
methods of CreDocument handle flow-less documents as well.