hide the progress bar if we're on the title/cover page
PBReader does this as well so PocketBook users expect this behaviour.

To be more precise, PBReader's logic is a bit more complex than this. It
hides the progress bar not just for the first (cover) page, but also
seems to hide it for the preamble content like copyright, table of
contents, etc. I suspect it just hides everything until the first page
that

    <guide><reference type="text" … /></guide>

refers to. This commit doesn't implement that, and hopefully people
won't request it.

Additionally, PBReader seems to remap the page numbers: since the first
page becomes 0 to hide the progress bar, and the last page stays the
same, the range has an extra page. The exact formula used is unknown,
but in practice it just ends up skipping a page number at the end of the
range (one can't do better than that if the input is an integer).

I believe it's better to skip a page number at the beginning, as there
actually exists a page that can be considered read—the cover page—we
just never show a progress bar saying that only one page is read.
Implementing this is also much easier.

Fixes: https://www.mobileread.com/forums/showthread.php?p=4373638#post4373638