Add "Scan home dir" entry to the Tools menu and action list
When we make changes to the file system — either by moving/copying in
the file manager, by deleting a book that's just been finished, or by
adding books using the Calibre plugin — the user expects PocketBook's
library and home screen to be updated.

This commit implements a function that asks the scanner task to rescan
the home directory (as set in KOReader's file manager). Unfortunately,
KOReader currently doesn't expose any hooks/events for the file system
being changed, so this function is only made available as a menu item.

The full functionality of automatically rescanning whenever necessary
shall be implemented later, when hooks/events in KOReader are available.

Related: https://github.com/ckilb/pocketbooksync.koplugin/issues/12